Allow {GCC,BINUTILS}_PLUGINS when STATIC_TOOLCHAIN
I am experimenting to see how badly it breaks here. It is not
recommended to use dlopen() from a static executable, because
you need to use the exact same glibc at runtime that was used
at buildtime.

crosstool-ng prevents doing this, but currently (as of about
four days ago) upstream actually would build some LTO stuff.
In particular, the stage2 GCC build the liblto.so plugin and
also gcc-ar, gcc-nm and gcc-ranlib.

I managed to get gcc-ar to run if I put liblto.so in the dir
that it is searched for.

A complete hack and probably not good enough. Instead we need
to implement MOSTLY_STATIC_TOOLCHAIN support where all the
companion libs are built statically with -fPIC.

We actually need to disallow GCC plugins here but it is possible
to take the plugin from the previous stage and use it. How well
this works in practice is not clear.